Create ExtensionHost component
